+++ ard/farm_aero/flowfarm/__init__.py
@@ -1,0 +1,10 @@
+from .component import FLOWFarmAEP, FLOWFarmBatchPower, FLOWFarmComponent
+from ._jl_bootstrap import ensure_flowfarm_loaded, get_julia_runtime
+
+__all__ = [
+    "FLOWFarmAEP",
+    "FLOWFarmBatchPower",
+    "FLOWFarmComponent",
+    "ensure_flowfarm_loaded",
+    "get_julia_runtime",
+]
