+++ ard/farm_aero/flowfarm/julia_env/Project.toml
@@ -3,7 +3,7 @@ version = "0.1.0"
 
 [deps]
 FLOWFarm = "eb2d4cfc-2064-11ea-0a1c-63d372e6a848"
-Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 
 [compat]
+FLOWFarm = "1"
 julia = "1.10"
