+++ ard/flowfarm/_jl_bootstrap.py
@@ -1,121 +1,5 @@
-# ard/farm_aero/flowfarm/_jl_bootstrap.py
-from __future__ import annotations
-import os
-import pathlib
-import warnings
+import sys
 
-# Hard-coded pin (change/remove later as needed)
-FLOWFARM_GIT_URL = "https://github.com/byuflowlab/FLOWFarm.jl"
-FLOWFARM_REV = "master"  # <-- BRANCH PIN
+from ard.farm_aero.flowfarm import _jl_bootstrap as _bootstrap
 
-_jl_module = None
-_jl_runtime = None
-_flowfarm_env_initialized = False
-
-
-def _normalize_juliacall_env_vars():
-    """Normalize JuliaCall env vars so Ard owns bootstrap by default.
-
-    Set ARD_FLOWFARM_RESPECT_JULIACALL_ENV=1 to keep user-provided overrides.
-    """
-    if os.environ.get("ARD_FLOWFARM_RESPECT_JULIACALL_ENV") == "1":
-        return
-
-    project = os.environ.get("PYTHON_JULIACALL_PROJECT")
-    exe = os.environ.get("PYTHON_JULIACALL_EXE")
-    if project or exe:
-        warnings.warn(
-            "Ignoring external JuliaCall overrides in Ard bootstrap. "
-            "Set ARD_FLOWFARM_RESPECT_JULIACALL_ENV=1 to keep user-provided "
-            "PYTHON_JULIACALL_PROJECT/PYTHON_JULIACALL_EXE values.",
-            UserWarning,
-            stacklevel=2,
-        )
-        os.environ.pop("PYTHON_JULIACALL_PROJECT", None)
-        os.environ.pop("PYTHON_JULIACALL_EXE", None)
-
-
-def get_julia_runtime():
-    """Return (Main, Pkg) from JuliaCall with Ard-safe bootstrap behavior."""
-    global _jl_runtime
-    if _jl_runtime is not None:
-        return _jl_runtime
-
-    _normalize_juliacall_env_vars()
-    from juliacall import Main as jl_main
-    from juliacall import Pkg as jl_pkg
-
-    _jl_runtime = (jl_main, jl_pkg)
-    return _jl_runtime
-
-
-def _is_manifest_mismatch_error(exc: Exception) -> bool:
-    text = str(exc)
-    markers = [
-        "different julia version",
-        "manifest generated by a different version of Julia",
-        "Could not locate the source code for the StyledStrings package",
-    ]
-    text_lower = text.lower()
-    return any(marker.lower() in text_lower for marker in markers)
-
-
-def _rebuild_flowfarm_env(jl_pkg, env_dir: pathlib.Path):
-    """Rebuild Ard FLOWFarm manifest for the currently running Julia runtime."""
-    manifest_path = env_dir / "Manifest.toml"
-    if manifest_path.exists():
-        manifest_path.unlink()
-
-    # Ensure FLOWFarm source is pinned and captured in the new manifest.
-    try:
-        jl_pkg.rm("FLOWFarm")
-    except Exception:
-        pass
-    jl_pkg.add(url=FLOWFARM_GIT_URL, rev=FLOWFARM_REV)
-    jl_pkg.resolve()
-    jl_pkg.instantiate()
-
-
-def ensure_flowfarm_loaded():
-    """Activate Ard Julia env and load FLOWFarm in Julia Main."""
-    global _flowfarm_env_initialized
-    jl_main, jl_pkg = get_julia_runtime()
-    if not _flowfarm_env_initialized:
-        env_dir = pathlib.Path(__file__).parent / "julia_env"
-        jl_pkg.activate(str(env_dir))
-        try:
-            jl_pkg.instantiate()
-        except Exception as exc:
-            if not _is_manifest_mismatch_error(exc):
-                raise
-            warnings.warn(
-                "FLOWFarm Julia manifest/runtime mismatch detected. Rebuilding "
-                "manifest for the current Julia runtime.",
-                UserWarning,
-                stacklevel=2,
-            )
-            _rebuild_flowfarm_env(jl_pkg, env_dir)
-        _flowfarm_env_initialized = True
-
-    if "FLOWFarm" not in dir(jl_main):
-        jl_main.seval("using FLOWFarm")
-    return jl_main
-
-
-def get_julia_module():
-    """Initialize Julia once, activate env, and return a private module with FLOWFarm loaded."""
-    global _jl_module
-    if _jl_module is not None:
-        return _jl_module
-
-    ensure_flowfarm_loaded()
-    import juliacall
-
-    jl = juliacall.newmodule(
-        "ArdFLOWFarm"
-    )  # recommended pattern to avoid polluting Main [1](https://juliapy.github.io/PythonCall.jl/stable/juliacall/)
-    jl.seval(
-        "using FLOWFarm"
-    )  # standard FLOWFarm usage/installation [2](https://github.com/byuflowlab/FlowFarm.jl)
-    _jl_module = jl
-    return _jl_module
+sys.modules[__name__] = _bootstrap
