+++ test/conftest.py
@@ -1,9 +1,4 @@
-import os
 from pathlib import Path
-
-
-# Disable OpenMDAO auto-report generation (n2/inputs/html artifacts) in tests.
-os.environ.setdefault("OPENMDAO_REPORTS", "0")
 
 
 def pytest_sessionfinish(session, exitstatus):
@@ -12,9 +7,7 @@ def pytest_sessionfinish(session, exitstatus):
     # remove pytest and OpenMDAO report output directories from cwd
     for pattern in ("pytest*_out", "__main__*_out"):
         for out_dir in Path().glob(pattern):
-            for root, dirs, files in out_dir.walk(
-            top_down=False
-        ):  # walk the directory
+            for root, dirs, files in out_dir.walk(top_down=False):  # walk the directory
                 for name in files:
                     (root / name).unlink()  # remove subdirectory files, and
                 for name in dirs:
