+++ test/flowfarm/unit/test_jl_bootstrap.py
@@ -1,30 +1,22 @@
 """
-Unit tests for ard/flowfarm/_jl_bootstrap.py.
+Unit tests for ard/farm_aero/flowfarm/_jl_bootstrap.py.
 
 juliacall is mocked entirely via sys.modules — Julia does not need to be installed
 for these tests.
 """
 
-import os
-import pathlib
 import sys
-import warnings
-from unittest.mock import MagicMock, call
+from unittest.mock import MagicMock
 
 import pytest
 
 import ard.flowfarm._jl_bootstrap as bootstrap
-
-# ---------------------------------------------------------------------------
-# Fixtures
-# ---------------------------------------------------------------------------
 
 
 @pytest.fixture(autouse=True)
 def reset_bootstrap_globals(monkeypatch):
     """Reset module-level singletons before each test so state never leaks."""
     monkeypatch.setattr(bootstrap, "_jl_runtime", None)
-    monkeypatch.setattr(bootstrap, "_jl_module", None)
     monkeypatch.setattr(bootstrap, "_flowfarm_env_initialized", False)
 
 
@@ -36,97 +28,6 @@ def mock_juliacall(monkeypatch):
     mock.Pkg = MagicMock(name="Pkg")
     monkeypatch.setitem(sys.modules, "juliacall", mock)
     return mock
-
-
-# ---------------------------------------------------------------------------
-# _normalize_juliacall_env_vars
-# ---------------------------------------------------------------------------
-
-
-class TestNormalizeJuliacallEnvVars:
-
-    def test_strips_project_and_exe_and_warns(self, monkeypatch):
-        monkeypatch.setenv("PYTHON_JULIACALL_PROJECT", "/some/project")
-        monkeypatch.setenv("PYTHON_JULIACALL_EXE", "julia")
-        monkeypatch.delenv("ARD_FLOWFARM_RESPECT_JULIACALL_ENV", raising=False)
-
-        with warnings.catch_warnings(record=True) as caught:
-            warnings.simplefilter("always")
-            bootstrap._normalize_juliacall_env_vars()
-
-        assert len(caught) == 1
-        assert "Ignoring" in str(caught[0].message)
-        assert "PYTHON_JULIACALL_PROJECT" not in os.environ
-        assert "PYTHON_JULIACALL_EXE" not in os.environ
-
-    def test_strips_only_project_when_exe_absent(self, monkeypatch):
-        monkeypatch.setenv("PYTHON_JULIACALL_PROJECT", "/some/project")
-        monkeypatch.delenv("PYTHON_JULIACALL_EXE", raising=False)
-        monkeypatch.delenv("ARD_FLOWFARM_RESPECT_JULIACALL_ENV", raising=False)
-
-        with warnings.catch_warnings(record=True) as caught:
-            warnings.simplefilter("always")
-            bootstrap._normalize_juliacall_env_vars()
-
-        assert len(caught) == 1
-        assert "PYTHON_JULIACALL_PROJECT" not in os.environ
-
-    def test_respects_opt_in_flag(self, monkeypatch):
-        monkeypatch.setenv("ARD_FLOWFARM_RESPECT_JULIACALL_ENV", "1")
-        monkeypatch.setenv("PYTHON_JULIACALL_EXE", "julia +1.10")
-
-        with warnings.catch_warnings(record=True) as caught:
-            warnings.simplefilter("always")
-            bootstrap._normalize_juliacall_env_vars()
-
-        assert len(caught) == 0
-        assert os.environ["PYTHON_JULIACALL_EXE"] == "julia +1.10"
-
-    def test_noop_when_no_vars_set(self, monkeypatch):
-        monkeypatch.delenv("PYTHON_JULIACALL_PROJECT", raising=False)
-        monkeypatch.delenv("PYTHON_JULIACALL_EXE", raising=False)
-        monkeypatch.delenv("ARD_FLOWFARM_RESPECT_JULIACALL_ENV", raising=False)
-
-        with warnings.catch_warnings(record=True) as caught:
-            warnings.simplefilter("always")
-            bootstrap._normalize_juliacall_env_vars()
-
-        assert len(caught) == 0
-
-
-# ---------------------------------------------------------------------------
-# _is_manifest_mismatch_error
-# ---------------------------------------------------------------------------
-
-
-class TestIsManifestMismatchError:
-
-    @pytest.mark.parametrize(
-        "msg",
-        [
-            "manifest resolved with a different julia version",
-            "Manifest generated by a different version of Julia",
-            "Could not locate the source code for the StyledStrings package",
-            "DIFFERENT JULIA VERSION",  # case-insensitive
-            "prefix text: different julia version :suffix",
-        ],
-    )
-    def test_returns_true_for_known_markers(self, msg):
-        assert bootstrap._is_manifest_mismatch_error(Exception(msg)) is True
-
-    def test_returns_false_for_unrelated_error(self):
-        assert (
-            bootstrap._is_manifest_mismatch_error(Exception("FLOWFarm not found"))
-            is False
-        )
-
-    def test_returns_false_for_empty_message(self):
-        assert bootstrap._is_manifest_mismatch_error(Exception("")) is False
-
-
-# ---------------------------------------------------------------------------
-# get_julia_runtime
-# ---------------------------------------------------------------------------
 
 
 class TestGetJuliaRuntime:
@@ -141,138 +42,26 @@ class TestGetJuliaRuntime:
         result2 = bootstrap.get_julia_runtime()
         assert result1 is result2
 
-    def test_juliacall_import_not_repeated(self, mock_juliacall):
-        bootstrap.get_julia_runtime()
-        bootstrap.get_julia_runtime()
-        # Runtime was cached; Main is still the same mock object
-        assert bootstrap._jl_runtime is not None
-        assert bootstrap._jl_runtime[0] is mock_juliacall.Main
-
-
-# ---------------------------------------------------------------------------
-# _rebuild_flowfarm_env
-# ---------------------------------------------------------------------------
-
-
-class TestRebuildFlowfarmEnv:
-
-    def test_deletes_existing_manifest(self, tmp_path):
-        manifest = tmp_path / "Manifest.toml"
-        manifest.write_text("old manifest content")
-        mock_pkg = MagicMock()
-
-        bootstrap._rebuild_flowfarm_env(mock_pkg, tmp_path)
-
-        assert not manifest.exists()
-
-    def test_no_error_when_manifest_absent(self, tmp_path):
-        mock_pkg = MagicMock()
-        bootstrap._rebuild_flowfarm_env(mock_pkg, tmp_path)  # should not raise
-
-    def test_adds_flowfarm_from_pinned_url(self, tmp_path):
-        mock_pkg = MagicMock()
-
-        bootstrap._rebuild_flowfarm_env(mock_pkg, tmp_path)
-
-        mock_pkg.add.assert_called_once_with(
-            url=bootstrap.FLOWFARM_GIT_URL,
-            rev=bootstrap.FLOWFARM_REV,
-        )
-
-    def test_calls_resolve_and_instantiate(self, tmp_path):
-        mock_pkg = MagicMock()
-
-        bootstrap._rebuild_flowfarm_env(mock_pkg, tmp_path)
-
-        mock_pkg.resolve.assert_called_once()
-        mock_pkg.instantiate.assert_called_once()
-
-    def test_rm_failure_is_silenced(self, tmp_path):
-        mock_pkg = MagicMock()
-        mock_pkg.rm.side_effect = RuntimeError("FLOWFarm not in environment")
-
-        bootstrap._rebuild_flowfarm_env(mock_pkg, tmp_path)
-
-        # add should still be called despite rm failing
-        mock_pkg.add.assert_called_once()
-
-
-# ---------------------------------------------------------------------------
-# ensure_flowfarm_loaded
-# ---------------------------------------------------------------------------
-
 
 class TestEnsureFlowfarmLoaded:
 
-    def test_activates_julia_env(self, mock_juliacall):
-        bootstrap.ensure_flowfarm_loaded()
-        mock_juliacall.Pkg.activate.assert_called_once()
-
-    def test_instantiates_julia_env(self, mock_juliacall):
-        bootstrap.ensure_flowfarm_loaded()
-        mock_juliacall.Pkg.instantiate.assert_called_once()
-
-    def test_calls_seval_to_load_flowfarm(self, mock_juliacall):
-        bootstrap.ensure_flowfarm_loaded()
-        mock_juliacall.Main.seval.assert_called_once_with("using FLOWFarm")
-
-    def test_returns_jl_main(self, mock_juliacall):
+    def test_activates_instantiates_and_loads_flowfarm(self, mock_juliacall):
         result = bootstrap.ensure_flowfarm_loaded()
+
+        mock_juliacall.Pkg.activate.assert_called_once()
+        mock_juliacall.Pkg.instantiate.assert_called_once()
+        mock_juliacall.Main.seval.assert_called_once_with("using FLOWFarm")
         assert result is mock_juliacall.Main
 
     def test_does_not_reinitialize_on_second_call(self, mock_juliacall):
         bootstrap.ensure_flowfarm_loaded()
+        mock_juliacall.Main.FLOWFarm = MagicMock(name="FLOWFarm")
         mock_juliacall.Pkg.activate.reset_mock()
         mock_juliacall.Pkg.instantiate.reset_mock()
+        mock_juliacall.Main.seval.reset_mock()
 
         bootstrap.ensure_flowfarm_loaded()
 
         mock_juliacall.Pkg.activate.assert_not_called()
         mock_juliacall.Pkg.instantiate.assert_not_called()
-
-    def test_rebuilds_manifest_on_version_mismatch(self, mock_juliacall, monkeypatch):
-        mock_juliacall.Pkg.instantiate.side_effect = Exception(
-            "manifest resolved with a different julia version"
-        )
-        rebuild_mock = MagicMock()
-        monkeypatch.setattr(bootstrap, "_rebuild_flowfarm_env", rebuild_mock)
-
-        with warnings.catch_warnings(record=True) as caught:
-            warnings.simplefilter("always")
-            bootstrap.ensure_flowfarm_loaded()
-
-        rebuild_mock.assert_called_once()
-        assert any("mismatch" in str(w.message).lower() for w in caught)
-
-    def test_reraises_non_mismatch_errors(self, mock_juliacall):
-        mock_juliacall.Pkg.instantiate.side_effect = RuntimeError("disk full")
-
-        with pytest.raises(RuntimeError, match="disk full"):
-            bootstrap.ensure_flowfarm_loaded()
-
-
-# ---------------------------------------------------------------------------
-# get_julia_module
-# ---------------------------------------------------------------------------
-
-
-class TestGetJuliaModule:
-
-    def test_returns_module_object(self, mock_juliacall, monkeypatch):
-        mock_juliacall.newmodule = MagicMock(return_value=MagicMock(name="ArdFLOWFarm"))
-        monkeypatch.setitem(sys.modules, "juliacall", mock_juliacall)
-
-        result = bootstrap.get_julia_module()
-
-        mock_juliacall.newmodule.assert_called_once_with("ArdFLOWFarm")
-        assert result is not None
-
-    def test_singleton_returns_same_module(self, mock_juliacall, monkeypatch):
-        mock_juliacall.newmodule = MagicMock(return_value=MagicMock(name="ArdFLOWFarm"))
-        monkeypatch.setitem(sys.modules, "juliacall", mock_juliacall)
-
-        result1 = bootstrap.get_julia_module()
-        result2 = bootstrap.get_julia_module()
-
-        assert result1 is result2
-        mock_juliacall.newmodule.assert_called_once()  # not called twice
+        mock_juliacall.Main.seval.assert_not_called()
